pyproject.toml
[tool.poetry]
name = "encoder"
version = "0.1.0"
description = ""
authors = ["Valentin LAMBOLEY <me@yubo.dev>"]
readme = "README.md"
packages = [
    { include = "encoder" },
]

[tool.poetry.scripts]
irilis-encoder = "encoder.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
pydantic = "^2.4.0"
qbittorrent-api = "^2023.9.53"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
paramiko = "^3.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
coverage = "^7.3.1"
darglint = "^1.8.1"
flake8 = "^6.1.0"
flake8-annotations = "^3.0.1"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.9.16"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.1.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pytype = "^2023.9.19"
safety = "^2.3.5"
ipython = "^8.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.coverage.paths]
source = ["nursery"]

[tool.coverage.run]
branch = true
source = ["nursery"]

[tool.coverage.report]
show_missing = true
fail_under = 100
omit = ["*/unittests/*"]
exclude_lines = ["pragma: no cover"]

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
length_sort = false
combine_star = false
use_parentheses = true
import_heading_stdlib = "Standard Library"
import_heading_firstparty = "First Party"
import_heading_thirdparty = "Third Party"
import_heading_localfolder = "Local"
force_sort_within_sections = false
from_first = false